requirements.txt
fastapi
pydantic
uvicorn
sendgrid
openai
pandas
streamlit
gspread
python-dotenv
